Extracted universal arm jtag+swd code from jlink probe.
